pyproject.toml
[project]
name = "codigos-qr1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "dearpygui>=2.0.0",
    "pillow>=11.1.0",
    "pip>=25.1.1",
    "pywin32==310",
    "pyzbar>=0.1.9",
    "qrcode>=8.1",
]
